+++ CHANGELOG.md
@@ -3,11 +3,18 @@
 This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.17.0](https://github.com/markphelps/flipt/releases/tag/v0.17.0) - 2020-07-10
 
 ### Added
 
 * Check for newer versions of Flipt on startup. Can be disabled by setting `meta.check_for_updates=false` in config. [https://github.com/markphelps/flipt/pull/311](https://github.com/markphelps/flipt/pull/311)
+* Ability to configure database connections [https://github.com/markphelps/flipt/pull/313](https://github.com/markphelps/flipt/pull/313)
+* Prometheus metrics around database connections [https://github.com/markphelps/flipt/pull/314](https://github.com/markphelps/flipt/pull/314)
+* OpenTracing/Jaeger support [https://github.com/markphelps/flipt/pull/315](https://github.com/markphelps/flipt/pull/315)
+
+### Changed
+
+* Update FQDN for cache metrics
 
 ## [v0.16.0](https://github.com/markphelps/flipt/releases/tag/v0.16.0) - 2020-06-29